+++ src/lib/Warp.NET.RenderImpl.Ui/Components/IconButton.cs
@@ -11,20 +11,20 @@ public class IconButton : Button
 	private readonly Texture _texture;
 	private readonly Vector2 _textureSize;
 
-	public IconButton(Bounds bounds, Action onClick, ButtonStyle buttonStyle, Texture texture)
+	public IconButton(IBounds bounds, Action onClick, ButtonStyle buttonStyle, Texture texture)
 		: base(bounds, onClick, buttonStyle)
 	{
 		_texture = texture;
 		_textureSize = new(texture.Width, texture.Height);
 	}
 
-	public override void Render(Vector2i<int> parentPosition)
+	public override void Render(Vector2i<int> scrollOffset)
 	{
-		base.Render(parentPosition);
+		base.Render(scrollOffset);
 
 		Vector2i<int> scale = Bounds.Size;
 		Vector2i<int> topLeft = Bounds.TopLeft;
 		Vector2i<int> center = topLeft + scale / 2;
-		RenderImplUiBase.Game.SpriteRenderer.Schedule(_textureSize, (parentPosition + center).ToVector2(), Depth + 2, _texture, IsDisabled ? Color.HalfTransparentWhite : Color.White);
+		RenderImplUiBase.Game.SpriteRenderer.Schedule(_textureSize, (scrollOffset + center).ToVector2(), Depth + 2, _texture, IsDisabled ? Color.HalfTransparentWhite : Color.White);
 	}
 }