+++ src/lib/Warp.NET.RenderImpl.Ui/Components/ScrollViewer.cs
@@ -6,11 +6,12 @@ namespace Warp.NET.RenderImpl.Ui.Components;
 public class ScrollViewer<TContent> : AbstractScrollViewer<ScrollViewer<TContent>, TContent>
 	where TContent : AbstractScrollContent<TContent, ScrollViewer<TContent>>, IScrollContent<TContent, ScrollViewer<TContent>>
 {
-	public ScrollViewer(Bounds bounds)
+	public ScrollViewer(IBounds bounds, int scrollbarWidth)
 		: base(bounds)
 	{
-		Content = TContent.Construct(bounds.CreateNested(0, 0, 0.95f, 1), this);
-		Scrollbar = new Scrollbar(Bounds.CreateNested(0.95f, 0, 0.05f, 1), SetScrollPercentage);
+		int division = bounds.Size.X - scrollbarWidth;
+		Content = TContent.Construct(bounds.CreateNested(0, 0, division, bounds.Size.Y), this);
+		Scrollbar = new Scrollbar(bounds.CreateNested(division, 0, scrollbarWidth, bounds.Size.Y), SetScrollPercentage);
 
 		NestingContext.Add(Content);
 		NestingContext.Add(Scrollbar);