+++ src/lib/Warp.NET.RenderImpl.Ui/Components/TextInput.cs
@@ -11,10 +11,8 @@ public class TextInput : AbstractTextInput
 {
 	private const float _cursorTimerSwitch = 0.45f;
 
-	private TextInputStyle _textInputStyle;
-
 	public TextInput(
-		Bounds bounds,
+		IBounds bounds,
 		bool isNumeric,
 		Action<string>? onEnter,
 		Action<string>? onDeselect,
@@ -22,30 +20,24 @@ public class TextInput : AbstractTextInput
 		TextInputStyle textInputStyle)
 		: base(bounds, isNumeric, onEnter, onDeselect, onInput)
 	{
-		_textInputStyle = textInputStyle;
+		TextInputStyle = textInputStyle;
 	}
 
-	public TextInputStyle TextInputStyle
-	{
-		get => _textInputStyle;
-		set
-		{
-			CharWidth = RenderImplUiBase.Game.GetFontRenderer(value.FontSize).Font.CharWidth;
-			TextRenderingHorizontalOffset = value.TextRenderingHorizontalOffset;
-			_textInputStyle = value;
-		}
-	}
+	public TextInputStyle TextInputStyle { get; set; }
 
-	public override void Render(Vector2i<int> parentPosition)
+	protected override int CharWidth => RenderImplUiBase.Game.GetFontRenderer(TextInputStyle.FontSize).Font.CharWidth;
+
+	protected override int TextRenderingHorizontalOffset => TextInputStyle.TextRenderingHorizontalOffset;
+
+	public override void Render(Vector2i<int> scrollOffset)
 	{
-		base.Render(parentPosition);
+		base.Render(scrollOffset);
 
 		int padding = (int)MathF.Round((Bounds.Y2 - Bounds.Y1) / 4f);
 		Vector2i<int> borderVec = new(TextInputStyle.BorderSize);
-		Vector2i<int> center = Bounds.TopLeft + Bounds.Size / 2;
 
-		RenderImplUiBase.Game.RectangleRenderer.Schedule(Bounds.Size, center + parentPosition, Depth, IsSelected ? TextInputStyle.ActiveBorderColor : TextInputStyle.BorderColor);
-		RenderImplUiBase.Game.RectangleRenderer.Schedule(Bounds.Size - borderVec, center + parentPosition, Depth + 1, Hover ? TextInputStyle.HoverBackgroundColor : TextInputStyle.BackgroundColor);
+		RenderImplUiBase.Game.RectangleRenderer.Schedule(Bounds.Size, Bounds.Center + scrollOffset, Depth, IsSelected ? TextInputStyle.ActiveBorderColor : TextInputStyle.BorderColor);
+		RenderImplUiBase.Game.RectangleRenderer.Schedule(Bounds.Size - borderVec, Bounds.Center + scrollOffset, Depth + 1, Hover ? TextInputStyle.HoverBackgroundColor : TextInputStyle.BackgroundColor);
 
 		MonoSpaceFontRenderer fontRenderer = RenderImplUiBase.Game.GetFontRenderer(TextInputStyle.FontSize);
 
@@ -56,12 +48,12 @@ public class TextInput : AbstractTextInput
 			int cursorSelectionStartX = Bounds.X1 + selectionStart * fontRenderer.Font.CharWidth + padding;
 
 			int cursorWidth = KeyboardInput.GetSelectionLength() * fontRenderer.Font.CharWidth + 1;
-			Vector2i<int> cursorPosition = parentPosition + new Vector2i<int>(cursorSelectionStartX + cursorWidth / 2, center.Y);
+			Vector2i<int> cursorPosition = scrollOffset + new Vector2i<int>(cursorSelectionStartX + cursorWidth / 2, Bounds.Center.Y);
 			RenderImplUiBase.Game.RectangleRenderer.Schedule(new(cursorWidth, Bounds.Size.Y - borderVec.Y), cursorPosition, Depth + 2, hasSelection ? TextInputStyle.SelectionColor : TextInputStyle.CursorColor);
 		}
 
 		Vector2i<int> position = new(Bounds.X1 + padding, Bounds.Y1 + padding);
 
-		fontRenderer.Schedule(Vector2i<int>.One, parentPosition + position, Depth + 3, TextInputStyle.TextColor, KeyboardInput.Value.ToString(), TextAlign.Left);
+		fontRenderer.Schedule(Vector2i<int>.One, scrollOffset + position, Depth + 3, TextInputStyle.TextColor, KeyboardInput.Value.ToString(), TextAlign.Left);
 	}
 }